Removed intermediary dirs, selectable compression algorithms
This commit introduces the possibility of selecting between gzip, xz and
zstd compression. If no option is specified gzip will be used as default
for backwards compatibility.

Backups and restores no longer need to create an intermediary directory
prior to archiving and compressing; on backup this is done all in one.
On restore an intermediary dir is still needed.